requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
PyJWT==2.9.0

